feat(resetPassword): refactor code
- update file names

[#162949036]
