--- src/components/Home.js
@@ -1,17 +1,0 @@
-import React from 'react';
-import '../App.scss';
-
-const HomeView = () =>
-  (
-    <div className="App">
-      <header className="App-header">
-        
-        <p>
-            Welcome to Author's Haven
-        </p>
-            The place of choice.
-      </header>
-    </div>
-  );
-
-export default HomeView;
