+++ src/components/Home/index.jsx
@@ -1,34 +1,17 @@
 import React from 'react';
-import PropTypes from 'prop-types';
-import logo from '../../logo.png';
 import '../../App.scss';
 
-const HomeComponent = ({ user }) => (
-  <div className="App">
-    <header className="App-header">
-      <img src={logo} className="App-logo" alt="logo" />
-      {(user.username) ? (
-        <div>
-          <p>
-            {`Hello, ${user.username}`}
-          </p>
-          welcome to Authors Haven
-        </div>
-) : (
-  <div>
-    <p>
-            Welcome to Authors Haven
-    </p>
-          The place to be.
-  </div>
-      )}
+const HomeView = () =>
+  (
+    <div className="App">
+      <header className="App-header">
+        
+        <p>
+            Welcome to Author's Haven
+        </p>
+            The place of choice.
+      </header>
+    </div>
+  );
 
-    </header>
-  </div>
-);
-
-HomeComponent.propTypes = {
-  user: PropTypes.shape({}).isRequired,
-};
-
-export default HomeComponent;
+export default HomeView;
