+++ src/components/Navigation/index.jsx
@@ -1,36 +1,32 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
-import { Menu, Container } from 'semantic-ui-react';
-import './navigation.scss';
+import { Menu } from 'semantic-ui-react';
 
-const NavigationComponent = () => (
+const NavigationView = () => (
   <div>
 
-    <Container>
-      <Menu pointing secondary>
-        <NavLink to="/">
-          <Menu.Item name="Home" active />
-        </NavLink>
-        <NavLink to="/articles">
-          <Menu.Item name="Articles" />
-        </NavLink>
-        <NavLink to="/profiles">
-          <Menu.Item name="Profile" />
-        </NavLink>
-        <Menu.Menu position="right">
-          <NavLink to="/login">
-            <Menu.Item
-              name="Login"
-            />
-          </NavLink>
-          <NavLink to="/register">
-            <Menu.Item name="Register" />
-          </NavLink>
-        </Menu.Menu>
-      </Menu>
-    </Container>
+    <Menu text>
+      <NavLink to="/">
+        <Menu.Item name="Home" />
+      </NavLink>
+
+      <NavLink to="/login">
+        <Menu.Item name="Login" />
+      </NavLink>
+      <NavLink to="/register">
+        <Menu.Item name="SignUp" />
+      </NavLink>
+      <NavLink to="/articles">
+        <Menu.Item name="Articles" />
+      </NavLink>
+      <NavLink to="/profiles">
+        <Menu.Item name="Profile" />
+      </NavLink>
+    </Menu>
 
   </div>
 
 );
-export default NavigationComponent;
+
+
+export default NavigationView;
