+++ src/components/newPassword/logo.png
