+++ src/components/newPassword/newpassword.scss
