+++ src/components/resetPassword/resetpassword.scss
