+++ src/views/HomeView/index.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import HomeView from '../../components/Home';
+
+class Home extends React.Component {
+  render() {
+    return (
+      <HomeView />
+    );
+  }
+}
+
+export default Home;
