+++ src/views/NavigationView/index.jsx
@@ -1,0 +1,12 @@
+import React,{Component} from 'react';
+import NavigationView from '../../components/Navigation';
+
+class Navigation extends Component{
+  render(){
+    return(
+      <NavigationView />
+    );
+  }
+}
+
+export default Navigation;
